Extracted entity structures
