+++ entities/entities_structures.h
@@ -1,9 +1,7 @@
-#ifndef ARKANOID_ENTITY_H
-#define ARKANOID_ENTITY_H
+#ifndef ARKANOID_ENTITIES_STRUCTURES_H
+#define ARKANOID_ENTITIES_STRUCTURES_H
 
-#include "config.h"
-#include "math/math.h"
-#include "textures.h"
+#include "../math/math.h"
 
 typedef struct AnimatedEntity AnimatedEntity;
 typedef struct Ball Ball;
@@ -46,10 +44,4 @@ struct VAUS {
     SpecificType active_powerup;
 };
 
-AnimatedEntity create_entity(SpecificType type, Point position);
-AnimatedEntity create_harmful(SpecificType harmfulType);
-AnimatedEntity create_capsule();
-Ball create_ball(Point position);
-VAUS create_VAUS(Point position);
-void update_VAUS_size(VAUS *vaus, int size);
-#endif // ARKANOID_ENTITY_H
+#endif //ARKANOID_ENTITIES_STRUCTURES_H
