+++ entities/entity.c
