+++ textures.h
@@ -1,6 +1,7 @@
 #ifndef ARKANOID_TEXTURES_H
 #define ARKANOID_TEXTURES_H
 
+#include "levels.h"
 #include "math/math.h"
 #include <SDL.h>
 #include <stdbool.h>
@@ -19,10 +20,13 @@ typedef enum {
 } Textures;
 
 void init_texture();
+
 void get_texture_dimensions(Textures texture, int *pos_x, int *pos_y,
                             int *width, int *height);
+
 void draw_texture(SDL_Surface *surface, Textures texture, int x, int y,
                   bool centered);
-void draw_vaus(SDL_Surface *surface, Point position, int size);
+
+void draw_vaus(SDL_Surface *surface, VAUS vaus);
 
 #endif // ARKANOID_TEXTURES_H
